replace $use_rvm with $install_method to enable more flexible installation possibilities
The $use_rvm variable was limited to either using rvm or not, so we replace that
with $install_method, defaulting to the previous usage

If you set $use_rvm it to not use rvm, then the module would just use the
gem. This made it so you couldn't install bundler via a package. So the
$install_method was added which enabled you to alter how the non-rvm
installation was provided.

Unfortunately, it was a mistake to have both $use_rvm and $install_method
because $use_rvm is 'true' by default, so if you tried to set an install_method,
then it wouldn't work because it would just use the rvm method. So in order to
install via something other than rvm or gem, you would need to do both use_rvm
=> false; install_method => <whatever>.

Just having the install_method parameter is much cleaner, because it is generic,
doesn't require multiple settings when not installing via rvm or gem, and it
defaults to what the module was doing before (using rvm by default).