fix $install_method to work properly
unfortunately, if one attempts to set the $install_method to an empty string,
puppet complains that you are interning an empty string. If you set it to undef,
then the parameterized default will get chosen.

To fix this, I replaced the $bundler::params::install_method default class
parameter value with 'rvm' (the value that $bundler::params::install_method
defaulted to) and then setup a test to determine if the $install_method was set
to undef, if so, then we set a second variable ($provider_method) to undef and
use that.